+++ example/dump_events.php
@@ -34,7 +34,7 @@ class MyEventSubscribers extends EventSubscribers
         echo $event;
 
         // all events got JsonSerializable implementation
-        //echo json_encode($event, JSON_PRETTY_PRINT);
+        echo json_encode($event, JSON_PRETTY_PRINT);
 
         echo 'Memory usage ' . round(memory_get_usage() / 1048576, 2) . ' MB' . PHP_EOL;
     }