+++ src/MySQLReplication/Event/DTO/EventDTO.php
@@ -45,13 +45,4 @@ abstract class EventDTO extends Event implements \JsonSerializable
      * @return string
      */
     abstract public function __toString();
-
-    /**
-     * Specify data which should be serialized to JSON
-     * @link http://php.net/manual/en/jsonserializable.jsonserialize.php
-     * @return mixed data which can be serialized by <b>json_encode</b>,
-     * which is a value of any type other than a resource.
-     * @since 5.4.0
-     */
-    abstract public function jsonSerialize();
 }