+++ src/MySQLReplication/Event/RotateEvent.php
@@ -36,7 +36,7 @@ class RotateEvent extends EventCommon
      */
     private function getSizeToRemoveByVersion()
     {
-        if (BinLogServerInfo::MYSQL_VERSION_MARIADB !== BinLogServerInfo::getVersion()) {
+        if (BinLogServerInfo::isMariaDb()) {
             return 8;
         }