+++ src/MySQLReplication/Exception/MySQLReplicationException.php
@@ -20,4 +20,29 @@ class MySQLReplicationException extends \Exception
 
     const UNKNOWN_JSON_TYPE_MESSAGE = 'Unknown JSON type: ';
     const UNKNOWN_JSON_TYPE_CODE = 300;
+
+    const USER_ERROR_MESSAGE = 'Incorrect user given';
+    const USER_ERROR_CODE = 400;
+    const IP_ERROR_MESSAGE = 'Incorrect IP given';
+    const IP_ERROR_CODE = 401;
+    const PORT_ERROR_MESSAGE = 'Incorrect port given should be numeric ';
+    const PORT_ERROR_CODE = 402;
+    const PASSWORD_ERROR_MESSAGE = 'Incorrect password type';
+    const PASSWORD_ERROR_CODE = 403;
+    const DB_NAME_ERROR_MESSAGE = 'Incorrect db name type';
+    const DB_NAME_ERROR_CODE = 404;
+    const CHARSET_ERROR_MESSAGE = 'Incorrect charset type';
+    const CHARSET_ERROR_CODE = 405;
+    const GTID_ERROR_MESSAGE = 'Incorrect gtid';
+    const GTID_ERROR_CODE = 406;
+    const SLAVE_ID_ERROR_MESSAGE = 'Incorrect slave id type';
+    const SLAVE_ID_ERROR_CODE = 407;
+    const BIN_LOG_FILE_NAME_ERROR_MESSAGE = 'Incorrect binlog name type';
+    const BIN_LOG_FILE_NAME_ERROR_CODE = 408;
+    const BIN_LOG_FILE_POSITION_ERROR_MESSAGE = 'Incorrect binlog position type';
+    const BIN_LOG_FILE_POSITION_ERROR_CODE = 409;
+    const MARIADBGTID_ERROR_MESSAGE = 'Maria gtid must be string';
+    const MARIADBGTID_ERROR_CODE = 410;
+    const TABLE_CACHE_SIZE_ERROR_MESSAGE = 'Table cache must be integer';
+    const TABLE_CACHE_SIZE_ERROR_CODE = 411;
 }